Fix parser test running 1-4 times
