+++ packages/compiler/test/parser-utils.test.ts
@@ -4,7 +4,7 @@ import { SyntaxKind, TypeSpecScriptNode } from "../src/core/index.js";
 import { getNodeAtPosition, parse } from "../src/core/parser.js";
 import { Node } from "../src/core/types.js";
 import { extractCursor } from "../src/testing/test-server-host.js";
-import { dumpAST } from "./parser.test.js";
+import { dumpAST } from "./ast-test-utils.js";
 
 describe("compiler: parser utils", () => {
   describe("getNodeAtPosition", () => {
